fix: network byte order, header packing and stream desync
- pack MessageHeader with pragma pack(1)
- serialize header id/length in network byte order
- fix receive header completeness check
- unify packet ID type to uint16_t
- split setsockopt calls in Acceptor
- loop send() until all bytes are sent

+++ examples/headers/common/Datatype.hpp
@@ -5,7 +5,7 @@
 #include <cstring>
 
 struct example_packet_data {
-    static constexpr uint32_t ID = 1;
+    static constexpr uint16_t ID = 1;
 
     int value;
